pyproject.toml
[project]
name = "student-manager"
version = "0.1.0"
description = "塾成績管理システム - FastAPI + HTMX"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "itsdangerous>=2.2.0",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "starlette>=0.41.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.27",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint]
select = ["E", "F", "I"]
